NewCustomerStreaksTest cleaned for Allure
